CORDA-3291 Remove hanging test and move tests into `KillFlowTest`
